Partial merge with pychrm. Adding transform algorithm registration and integration with FeatureNames
git-svn-id: http://wnd-charm.googlecode.com/svn/wndchrm/trunk@588 d860a611-dd6a-fbde-3e4e-7fa7191099b5

+++ gsl/specfunc.cpp
@@ -993,7 +993,7 @@ int gsl_sf_psi_1_e(const double x, gsl_sf_result * result)
   else if(x > -5.0)
   {
     /* Abramowitz + Stegun 6.4.6 */
-    int M = -floor(x);
+    int M = (int)-floor(x);
     double fx = x + M;
     double sum = 0.0;
     int m;
